Editando migrate

+++ sigo/db/migrate/20140523131948_remove_ancient_tables_from_database.rb
@@ -1,8 +1,16 @@
 class RemoveAncientTablesFromDatabase < ActiveRecord::Migration
   def change
-    drop_table :attachment_arbitration_documents
-    drop_table :attachment_contract_documents
+  	if AttachmentArbitrationDocuments.table_exists?
+      drop_table :attachment_arbitration_documents
+  	end
+  	if AttachmentContractDocuments.table_exists?
+      drop_table :attachment_contract_documents
+  	end
+  	if AttachmentIntegratedBases_documents.table_exists?
     drop_table :attachment_integrated_bases_documents
-    drop_table :attachment_testimony_of_consortium_documents
+	end
+	if AttachmentTestimonyOfConsortiumDocuments.table_exists?
+      drop_table :attachment_testimony_of_consortium_documents
+    end
   end
 end
